Create cells.rs, refactor layout.rs
