+++ crates/typst/src/layout/grid/mod.rs
@@ -1,11 +1,12 @@
+mod cells;
 mod layout;
 mod lines;
 mod rowspans;
 
-pub use self::layout::{
-    Cell, CellGrid, Celled, GridLayouter, ResolvableCell, ResolvableGridChild,
-    ResolvableGridItem,
+pub use self::cells::{
+    Cell, CellGrid, Celled, ResolvableCell, ResolvableGridChild, ResolvableGridItem,
 };
+pub use self::layout::GridLayouter;
 pub use self::lines::LinePosition;
 
 use std::num::NonZeroUsize;
